Revert "[macsec] Add a pytest option --enable_macsec (#5362)"
This reverts commit 4321596c7c7b1554e2ae053ef74ac2652651a525.
