+++ tests/conftest.py
@@ -63,7 +63,6 @@ pytest_plugins = ('tests.common.plugins.ptfadapter',
                   'tests.common.dualtor',
                   'tests.vxlan',
                   'tests.decap',
-                  'tests.macsec',
                   'tests.common.plugins.allure_server',
                   'tests.common.plugins.conditional_mark')
 
@@ -403,7 +402,7 @@ def k8scluster(k8smasters):
     return k8s_master_cluster
 
 
-@pytest.fixture(scope="session")
+@pytest.fixture(scope="module")
 def nbrhosts(ansible_adhoc, tbinfo, creds, request):
     """
     Shortcut fixture for getting VM host
@@ -602,8 +601,9 @@ def creds_on_dut(duthost):
 
     return creds
 
-@pytest.fixture(scope="session")
-def creds(duthost):
+@pytest.fixture(scope="module")
+def creds(duthosts, rand_one_dut_hostname):
+    duthost = duthosts[rand_one_dut_hostname]
     return creds_on_dut(duthost)
