+++ tests/macsec/macsec_platform_helper.py
@@ -1,4 +1,5 @@
 import ast
+import collections
 import re
 from multiprocessing.pool import ThreadPool
 
@@ -16,6 +17,28 @@ def global_cmd(duthost, nbrhosts, cmd):
         pool.apply_async(nbr["host"].command, args=(cmd, ))
     pool.close()
     pool.join()
+
+
+def find_links(duthost, tbinfo, filter):
+    mg_facts = duthost.get_extended_minigraph_facts(tbinfo)
+    for interface, neighbor in mg_facts["minigraph_neighbors"].items():
+        filter(interface, neighbor, mg_facts, tbinfo)
+
+
+def find_links_from_nbr(duthost, tbinfo, nbrhosts):
+    links = collections.defaultdict(dict)
+
+    def filter(interface, neighbor, mg_facts, tbinfo):
+        if neighbor["name"] not in nbrhosts.keys():
+            return
+        port = mg_facts["minigraph_neighbors"][interface]["port"]
+        links[interface] = {
+            "name": neighbor["name"],
+            "host": nbrhosts[neighbor["name"]]["host"],
+            "port": port
+        }
+    find_links(duthost, tbinfo, filter)
+    return links
 
 
 def sonic_db_cli(host, cmd):
